Restrict bot hole planting to the host
Add a check for bot hole planting to disable it for players other than the host. Specify as an override to ensure that the package overrides the function's definition.

This is specified as host-only as it would be difficult to reliably check for admin if the hole is being loaded while its owner is offline.

See issue #25.