Rework lot name formatting
Overhaul the way lot names are formatted, changing lots to store three parameters: The server host, the lot's owner BLID, and the lot's ID. Separate parameters using underscores in the brick names. Parse this information on init of existing lots to determine if the save is from another server.

If the lot is from another server, display a prominent warning, along with a placeholder function that is intended to later convert lots based on the stored ID in their NT Object name.