Clean up lot registry and replace lot saver
Replace the saver used by the lot registry with McTwist's saver. Leave most of the calls unchanged, but remove bloat in the form of unreachable errors and erroneous tests.

Update lot registry saving so that it saves at the same time as the regular Sassy saver. This will prevent inconsistencies between purchased lots and user cash if a hard crash occurs.