Implement City object for obtaining data
Implement a new generic object for obtaining player data, an object named City with ::get and ::set functions for obtaining respective player data. This new object is intended to be used in place of CityRPGData calls from this point forward.

Utilize this new object in the admin mode toggler as a test.